+++ src/lib/Components/Inbox/Conversations/Message.tsx
@@ -4,6 +4,7 @@ import * as Relay from "react-relay"
 
 import { BodyText, MetadataText, SmallHeadline } from "../Typography"
 
+import Avatar from "./Avatar"
 import ImagePreview from "./Preview/Attachment/ImagePreview"
 import PDFPreview from "./Preview/Attachment/PDFPreview"
 
@@ -28,13 +29,6 @@ const Container = styled(HorizontalLayout)`
 
 `
 
-const Avatar = styled.View`
-  height: 20
-  width: 20
-  borderRadius: 20
-  backgroundColor: ${colors["gray-regular"]}
-`
-
 const Header = styled(HorizontalLayout)`
   alignSelf: stretch
   marginBottom: 10
@@ -56,7 +50,9 @@ const PDFPreviewContainer = styled.View`marginBottom: 10;`
 
 interface Props extends RelayProps {
   senderName: string
+  initials?: string
   artworkPreview?: JSX.Element
+  relay?: Relay.RelayProp
 }
 
 export class Message extends React.Component<Props, any> {
@@ -85,29 +81,31 @@ export class Message extends React.Component<Props, any> {
   }
 
   render() {
-    const date = moment(this.props.message.created_at).fromNow(true)
+    const { artworkPreview, initials, message, senderName } = this.props
+    const isSent = this.props.relay ? !this.props.relay.hasOptimisticUpdate(message) : true
 
     return (
       <Container>
-        <Avatar />
+        <Avatar isUser={message.is_from_user} initials={initials} />
         <TextContainer>
           <Header>
             <SenderName>
-              {this.props.senderName}
+              {senderName}
             </SenderName>
-            <MetadataText>
-              {date}
-            </MetadataText>
+            {isSent &&
+              <MetadataText>
+                {moment(message.created_at).fromNow(true)}
+              </MetadataText>}
           </Header>
-          {this.props.artworkPreview &&
+          {artworkPreview &&
             <ArtworkPreviewContainer>
-              {this.props.artworkPreview}
+              {artworkPreview}
             </ArtworkPreviewContainer>}
 
-          {this.renderAttachmentPreviews(this.props.message.attachments)}
+          {this.renderAttachmentPreviews(message.attachments)}
 
-          <BodyText>
-            {this.props.message.raw_text.split("\n\nAbout")[0]}
+          <BodyText disabled={!isSent}>
+            {message.raw_text.split("\n\nAbout")[0]}
           </BodyText>
         </TextContainer>
       </Container>
@@ -118,14 +116,15 @@ export class Message extends React.Component<Props, any> {
 export default Relay.createContainer(Message, {
   fragments: {
     message: () => Relay.QL`
-      fragment on MessageType {
+      fragment on Message {
         raw_text
         created_at
         is_from_user
         attachments {
+          id
           content_type
           download_url
-          id
+          file_name
           ${ImagePreview.getFragment("attachment")}
           ${PDFPreview.getFragment("attachment")}
         }