+++ src/lib/Containers/__tests__/Conversation-tests.tsx
@@ -10,13 +10,15 @@ it("looks correct when rendered", () => {
 })
 
 const props = {
+  initials: "JC",
   conversation: {
     id: "420",
     from: {
       name: "Anita Garibaldi",
       email: "anita@garibaldi.br",
+      initials: "AG",
     },
-    to: { name: "Kimberly Klark" },
+    to: { name: "Kimberly Klark", initials: "KK" },
     initial_message: "Adoro! Por favor envie-me mais informações",
     messages: {
       edges: [